chore: move makeId to utils
